+++ src/main/java/com/itquasar/multiverse/mail/message/Email.java
@@ -3,10 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.message;
 
-import com.itquasar.multiverse.mail.content.Content;
-import com.itquasar.multiverse.mail.envelope.Envelope;
+import com.itquasar.multiverse.mail.message.content.Content;
+import com.itquasar.multiverse.mail.message.envelope.Envelope;
 import javax.mail.Message;
 import javax.mail.Session;
