+++ src/main/java/com/itquasar/multiverse/mail/message/envelope/ImmutableEnvelope.java
@@ -1,4 +1,4 @@
-package com.itquasar.multiverse.mail.envelope;
+package com.itquasar.multiverse.mail.message.envelope;
 
 import com.itquasar.multiverse.mail.contact.Contact;
 import com.itquasar.multiverse.mail.contact.ImmutableRecipients;
