+++ src/main/java/com/itquasar/multiverse/mail/message/envelope/LazyEnvelope.java
@@ -1,4 +1,4 @@
-package com.itquasar.multiverse.mail.envelope;
+package com.itquasar.multiverse.mail.message.envelope;
 
 import com.itquasar.multiverse.mail.contact.Recipients;
 import com.itquasar.multiverse.mail.contact.Senders;
