+++ src/main/java/com/itquasar/multiverse/mail/part/MessagePart.java
@@ -1,5 +1,6 @@
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.part;
 
+import com.itquasar.multiverse.mail.message.ImmutableEmail;
 import com.itquasar.multiverse.mail.part.Disposition;
 import com.itquasar.multiverse.mail.part.GenericPart;
