+++ src/main/java/com/itquasar/multiverse/mail/server/Protocol.java
@@ -1,4 +1,4 @@
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.server;
 
 /**
  *
